Add @Nullable annotation to histogram to make it work with the NullAway library
A user pointed out that there's no way currently to use the default buckets (specified via `null`) in a project that
uses https://github.com/uber/NullAway. That library assumes that anything unannotated is @NotNull, which is aggressive but probably
the right move. To remedy this, I've just gone and explicitly marked `histogram` as nullable.

We can annotate further as we get time/touch the code naturally.

A note r.e. the findbugs dependency--this seems like a fairly standard source for the annotation, but if there's a better one let me know.