Add TestScope
This commit plucks select changes from uber-java/tally#69 and uber-java/tally#43
to bring uber-java/tally on par with uber-go/tally.

It adds a `TestScope` class which allows users of the library to get a
`Snapshot` of all the metrics emitted by the scope. Users can then assert if
certain metrics were created.